moved files and cleaned up some unused code
