+++ src/Components/ButtonGroup/ButtonGroup.js
@@ -6,17 +6,16 @@ import arrowBlack from "./wide-arrow-black.svg";
 
 // const links = [{ text, url }]
 const ButtonGroup = ({ links }) => (
-	<nav className="ButtonGroup">
-		{links.map((link, index) => (
-			<Link key={link.text} className="ButtonGroup-link" to={link.url}>
-				<span className="ButtonGroup-text">{link.text}</span>
-
-				<div className="ButtonGroup-arrow">
-					<img alt="arrow" src={index === 0 ? arrowBlack : arrowWhite} />
-				</div>
-			</Link>
-		))}
-	</nav>
+  <nav className="ButtonGroup">
+    {links.map((link, index) => (
+      <Link key={link.text} className="ButtonGroup-link" to={link.url}>
+        <span className="ButtonGroup-text">{link.text}</span>
+        <div className="ButtonGroup-arrow">
+          <img alt="arrow" src={index === 0 ? arrowBlack : arrowWhite} />
+        </div>
+      </Link>
+    ))}
+  </nav>
 );
 
 export default ButtonGroup;
