+++ src/Components/WorkoutForm/WorkoutForm.css
@@ -12,10 +12,6 @@
   justify-content: space-between;
 }
 
-.submit-buttons {
-
-}
-
 .form-button {
   display: block;
   width: 100%;
