+++ src/Components/WorkoutForm/WorkoutForm.js
@@ -1,7 +1,7 @@
 import { useState } from "react";
 import "./WorkoutForm.css";
-import SSButton from "./SSButton";
-import SSInput from "./SSInput";
+import SSButton from "../SSButton/SSButton";
+import SSInput from "../SSInput/SSInput";
 // Todo
 // 1. set weight to localState variable called updatedWeight
 // 2. function to change value of updatedWieght
